--- a/cardano-chain-gen/bench/benchfiles/pgpass-bench
+++ b/cardano-chain-gen/bench/benchfiles/pgpass-bench
@@ -1,0 +1,1 @@
+/var/run/postgresql:5432:bench:*:*