--- a/cardano-chain-gen/src/Cardano/Mock/Forging/Interpreter.hs
+++ b/cardano-chain-gen/src/Cardano/Mock/Forging/Interpreter.hs
@@ -18,6 +18,7 @@ module Cardano.Mock.Forging.Interpreter
   , NodeId (..)
   , initInterpreter
   , withInterpreter
+  , finalizeFingerprint
   , forgeNextFindLeader
   , forgeNext
   , forgeNextAfter