@kderme 2022-02-28
A more general caching approach is here https://github.com/input-output-hk/cardano-db-sync/pull/1074